+++ src/main/java/nl/tudelft/goalkeeper/parser/results/files/module/ModuleFile.java
@@ -1,7 +1,6 @@
 package nl.tudelft.goalkeeper.parser.results.files.module;
 
 import nl.tudelft.goalkeeper.parser.results.files.File;
-import nl.tudelft.goalkeeper.parser.results.files.module.parsers.Module;
 
 import java.io.IOException;
 import java.util.ArrayList;
